\name{addPitchCount}
\alias{addPitchCount}
\title{Add columns with relevant pitch count to the 'pitch' data frame.}
\usage{
  addPitchCount(df)
}
\arguments{
  \item{df}{data frame with all "pitch" attributes from
  "~/inning/inning_all.xml" files.}
}
\value{
  returns a vector with the proper pitch count for each
  record (ie, pitch) in \code{df}
}
\description{
  Add columns with relevant pitch count to the 'pitch' data
  frame.
}

