\name{createAtbatID}
\alias{createAtbatID}
\title{Assign each pitch an atbat ID}
\usage{
  createAtbatID(nodes)
}
\arguments{
  \item{nodes}{XML nodes from a set of URLs. These nodes
  should be from the "atbat" node.}
}
\value{
  returns a vector contains an atbat ID for each pitch
  thrown
}
\description{
  Assign each pitch an atbat ID
}

