\name{createInnings}
\alias{createInnings}
\title{Create columns to match an atbat with a inning (and side of that inning)}
\usage{
  createInnings(atbats)
}
\arguments{
  \item{atbats}{data frame with attributes from the atbat
  node}
}
\value{
  returns the original data frame with two additional
  columns
}
\description{
  Create columns to match an atbat with a inning (and side
  of that inning)
}

