\docType{data}
\name{fields}
\alias{fields}
\title{Field names for tables obtained by pitchRx::scrapeFX}
\description{
  A list of character vectors that specify the desired
  subset variables of the atbat, pitch, game and player
  tables. These vectors can be used to reduce parsing time
  of number of columns returned by \code{scrapeFX}.
}
\author{
  Carson Sievert
}
\keyword{datasets}

