% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gids}
\alias{gids}
\title{All MLB Gameday IDs from 2008-2013}
\format{A character vector}
\usage{
gids
}
\description{
A character vector with every "gameday" attribute in the
"game" element taken from scoreboard files like this one:
\url{http://gd2.mlb.com/components/game/mlb/year_2011/month_04/day_04/gid_2011_04_04_minmlb_nyamlb_1/miniscoreboard.xml}
Note they are ordered from oldest game to newest game.
}
\keyword{datasets}

