% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nonMLBgids}
\alias{nonMLBgids}
\title{All non-MLB Gameday IDs from 2008-2013}
\format{
A character vector
}
\usage{
nonMLBgids
}
\description{
A character vector with every "gameday" attribute in the "game" element taken from scoreboard files like this one:
\url{http://gd2.mlb.com/components/game/aaa/year_2013/month_06/day_08/gid_2013_06_08_freaaa_slcaaa_1/miniscoreboard.xml}
Note they are ordered from oldest game to newest game.
}
\keyword{datasets}
