% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{parseObs}
\alias{parseObs}
\title{parseObs}
\usage{
parseObs(obs = obs, innings_all = TRUE)
}
\arguments{
\item{obs}{An \code{XML2R} object from the \code{scrape} function.}

\item{innings_all}{A logical statement. Should an "innings_all" file be parsed, or an individual inning? The default is TRUE.}
}
