% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{pitchRx}
\alias{pitchRx}
\alias{pitchRx-package}
\title{PITCHf/x package}
\description{
PITCHf/x package
}
\author{
Carson Sievert
}
\seealso{
\url{http://cpsievert.github.com/pitchRx}
}

