\name{strikeFX}
\alias{strikeFX}
\title{Visualize PITCHf/x strikezones}
\usage{
  strikeFX(data, geom = "point", layer = list())
}
\arguments{
  \item{data}{PITCHf/x data to be visualized.}

  \item{geom}{type of geometry used for plotting.}

  \item{layer}{list of other ggplot2 (layered)
  modifications.}
}
\value{
  Returns a ggplot2 object.
}
\description{
  Pitch locations as they crossed home plate.
}
\details{
  Scatterplot with "px" on the horizonatl axis and "pz" on
  the vertical axis.
}
\examples{
#value
}

