\name{subsetFX}
\alias{subsetFX}
\title{Special subset handling for density estimates}
\usage{
  subsetFX(data, density)
}
\arguments{
  \item{data}{PITCHf/x data}

  \item{density}{either density1 or density2 passed on from
  \link{strikeFX}}
}
\description{
  Special subset handling for density estimates
}

