\name{updatePlayers}
\alias{updatePlayers}
\title{Update \code{players} data frame}
\usage{
  updatePlayers(new.urls)
}
\arguments{
  \item{new.urls}{new player urls added to the \code{urls}
  data frame}
}
\value{
  returns new player information in a data frame
}
\description{
  This function takes on (new) player urls and scrapes
  personal information on each player.
}

