\name{addPitchCount}
\alias{addPitchCount}
\title{Add columns with relevant pitch count to the 'pitch' data frame.}
\usage{
  addPitchCount(df)
}
\arguments{
  \item{df}{'pitch' data frame with attributes from
  "~/inning/inning_all.xml" files.}
}
\value{
  returns the original data frame with the proper pitch
  count columns attached at the end.
}
\description{
  Add columns with relevant pitch count to the 'pitch' data
  frame.
}

