% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{fields}
\alias{fields}
\title{Master list of tables and fields returned by \code{scrape}}
\description{
This data object is used to find to set of fields for
tables in a database. Since it's much easier to write to a
table with more fields (than vice versa), this object
contains every possible field for each table.
}
\author{
Carson Sievert
}

