% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{fields}
\alias{fields}
\title{Master list of tables and fields returned by \code{scrape}}
\format{A list of character vectors.}
\usage{
fields
}
\description{
This data object is as a template for fields and fields
types for each table. Since it's much easier to write to a
table with more fields (than vice versa), this object
contains every possible field for each table.
}
\keyword{datasets}

