% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSnapshots.R
\name{getSnapshots}
\alias{getSnapshots}
\title{Produce time sequenced pitch locations from PITCHf/x parameters}
\usage{
getSnapshots(data, interval = 0.01)
}
\arguments{
\item{data}{The nine PITCHf/x parameters used to determine the location of a pitch at a given time.}

\item{interval}{the amount of time between 'snapshots'}
}
\value{
Return a three dimensional array. The third dimension corresponds to different 'snapshots' of locations.
}
\description{
This function generates the x, y and z locations used in \link{animateFX} and \link{interactiveFX}.
}
\references{
http://baseball.physics.illinois.edu/KaganPitchfx.pdf
}
