\docType{data}
\name{gids}
\alias{gids}
\title{All Gameday IDs from 2008-2013}
\description{
  A character vector with every "gameday" attribute in the
  "game" element taken from scoreboard files like this one:
  \url{http://gd2.mlb.com/components/game/mlb/year_2011/month_04/day_04/gid_2011_04_04_minmlb_nyamlb_1/miniscoreboard.xml}
}
\author{
  Carson Sievert
}
\keyword{datasets}

